refactor: edit channels paths
