+++ package.json
@@ -10,7 +10,7 @@
   "husky": {
     "hooks": {
       "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
-      "pre-commit": "npm run format"
+      "pre-commit": "npm run lint"
     }
   },
   "release-it": {
