+++ src/services/guilds/index.ts
@@ -6,6 +6,8 @@ import { getByIdGuildsRouter } from './:guildId/get'
 import { getPublicDiscoverGuildsRouter } from './public/discover/get'
 import { postGuildsRouter } from './post'
 import { putByIdGuildsRouter } from './:guildId/put'
+import { getChannelsRouter } from './:guildId/channels/get'
+import { postChannelsRouter } from './:guildId/channels/post'
 
 const guildsRouter = Router()
 
@@ -15,5 +17,7 @@ guildsRouter.use('/', getByIdGuildsRouter)
 guildsRouter.use('/', getPublicDiscoverGuildsRouter)
 guildsRouter.use('/', deleteByIdGuildsRouter)
 guildsRouter.use('/', putByIdGuildsRouter)
+guildsRouter.use('/', getChannelsRouter)
+guildsRouter.use('/', postChannelsRouter)
 
 export { guildsRouter }
